Separate actions into multiple files
[MAILPOET-2682]
