+++ assets/js/src/settings/store/actions/index.ts
@@ -1,0 +1,5 @@
+export * from './settings';
+export * from './mss_and_premium';
+export { default as reinstall } from './reinstall';
+export { default as sendTestEmail } from './send_test_email';
+export { default as openWoocommerceCustomizer } from './open_woocommerce_customizer';
