+++ assets/js/src/settings/store/actions/open_woocommerce_customizer.ts
@@ -1,0 +1,17 @@
+export default function* openWoocommerceCustomizer(newsletterId?: string) {
+  let id = newsletterId;
+  if (!id) {
+    const { res, success, error } = yield {
+      type: 'CALL_API',
+      endpoint: 'settings',
+      action: 'set',
+      data: { 'woocommerce.use_mailpoet_editor': 1 },
+    };
+    if (!success) {
+      return { type: 'SAVE_FAILED', error };
+    }
+    id = res.data.woocommerce.transactional_email_id;
+  }
+  window.location.href = `?page=mailpoet-newsletter-editor&id=${id}`;
+  return null;
+}
