+++ assets/js/src/settings/store/actions/reinstall.ts
@@ -1,0 +1,16 @@
+import MailPoet from 'mailpoet';
+
+export default function* reinstall() {
+  MailPoet.Modal.loading(true);
+  const { success, error } = yield {
+    type: 'CALL_API',
+    endpoint: 'setup',
+    action: 'reset',
+  };
+  MailPoet.Modal.loading(false);
+  if (!success) {
+    return { type: 'SAVE_FAILED', error };
+  }
+  yield { type: 'TRACK_REINSTALLED' };
+  return { type: 'SAVE_DONE' };
+}
